requirements.txt
pandas==2.2.3
google-cloud-bigquery==3.30.0
google-cloud-bigquery-storage==2.29.1
db-dtypes==1.4.2
